# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoProxy API
  version: 1.0.0
  description: |
    Management API for GoProxy. The dashboard and API keys share this surface.

    Use **Authorize** with `Authorization: Bearer gpk_...`, or log in first
    (`POST /auth/login`) so the session cookie is sent with Try it out.

    Scopes are `{resource}:read` or `{resource}:write` for
    frontends, backends, acls, certs, users, apikeys, audit, stats, settings.

    IDs are optional on create. Omit them and GoProxy generates a slug.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: goproxy_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key, for example `gpk_...`
  parameters:
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string }
      example: app.example.com
...
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: bind is required }
    Deleted:
      description: Deleted
  schemas:
...
    StatusOK:
      type: object
      properties:
        status: { type: string, example: ok }
        version: { type: string, example: 1.0.0 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, format: password, example: change-me }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, format: password, example: old-password }
        new_password: { type: string, format: password, example: new-password }
    LoginResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        user: { $ref: '#/components/schemas/User' }
...
    Frontend:
      type: object
      required: [bind]
      properties:
        id: { type: string, example: https-443, description: Optional on create. Generated from bind and TLS. }
        name: { type: string, example: Public HTTPS }
        bind: { type: string, example: 0.0.0.0:443 }
        tls: { $ref: '#/components/schemas/FrontendTLS' }
        default: { $ref: '#/components/schemas/FrontendDefault' }
      example:
        name: Public HTTPS
        bind: 0.0.0.0:443
        tls:
          enabled: true
          min_version: "1.2"
          hsts: false
        default:
          https_redirect: false
...
    ACL:
      type: object
      required: [frontend, match, mode]
      properties:
        id: { type: string, example: app.example.com, description: Optional on create. Generated from the host. }
        name: { type: string, example: App }
        frontend: { type: string, example: https-443 }
        match: { $ref: '#/components/schemas/Match' }
        mode: { type: string, enum: [terminate, passthrough], example: terminate }
        backend: { type: string, example: app, description: Required unless this route only redirects }
        certificate: { type: string, example: app.example.com }
        middleware: { $ref: '#/components/schemas/Middleware' }
        rate_limit: { $ref: '#/components/schemas/RateLimit' }
      example:
        name: App
        frontend: https-443
        match: { host: app.example.com }
        mode: terminate
        backend: app
        certificate: app.example.com
        middleware:
          redirect_keep_path: true
...
    Backend:
      type: object
      required: [servers]
      properties:
        id: { type: string, example: app, description: Optional on create. Generated from the first server host. }
        name: { type: string, example: App }
        algorithm: { type: string, enum: [round_robin, least_conn, weighted, ip_hash], example: round_robin }
        servers:
          type: array
          items: { $ref: '#/components/schemas/BackendServer' }
        health: { $ref: '#/components/schemas/HealthCheck' }
        rate_limit: { $ref: '#/components/schemas/RateLimit' }
      example:
        name: App
        algorithm: round_robin
        servers:
          - url: http://127.0.0.1:3000
            role: primary
            weight: 1
        health:
          type: http
          path: /health
          interval: 5s
          timeout: 2s
          healthy_threshold: 2
          unhealthy_threshold: 3
...
    CertUpload:
      type: object
      required: [cert_pem, key_pem]
      properties:
        cert_pem: { type: string, example: "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n" }
        key_pem: { type: string, example: "-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n" }
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      description: Sets the `goproxy_session` cookie. Use this before Try it out on other endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example: { username: admin, password: change-me }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Error' }
...
  /auth/password:
    post:
      tags: [Auth]
      summary: Change own password
      description: Session users only. API keys cannot change a password. Requires the current password.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
            example:
              current_password: old-password
              new_password: new-password
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOK' }
        '400': { $ref: '#/components/responses/Error' }
  /frontends:
    get:
      tags: [Frontends]
      summary: List frontends
      responses:
        '200':
          description: Listeners
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Frontend' }
    post:
      tags: [Frontends]
      summary: Create frontend
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Frontend' }
            example:
              name: Public HTTPS
              bind: 0.0.0.0:443
              tls: { enabled: true, min_version: "1.2", hsts: false }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Frontend' }
        '400': { $ref: '#/components/responses/Error' }
  /frontends/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    put:
      tags: [Frontends]
      summary: Update frontend
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Frontend' }
            example:
              id: https-443
              name: Public HTTPS
              bind: 0.0.0.0:443
              tls: { enabled: true, min_version: "1.2", hsts: false }
              default: { https_redirect: false }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Frontend' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Frontends]
      summary: Delete frontend
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
  /acls:
    get:
      tags: [Routes]
      summary: List routes
      responses:
        '200':
          description: Host routes (ACLs)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ACL' }
    post:
      tags: [Routes]
      summary: Create route
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ACL' }
            example:
              name: App
              frontend: https-443
              match: { host: app.example.com }
              mode: terminate
              backend: app
              certificate: app.example.com
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ACL' }
        '400': { $ref: '#/components/responses/Error' }
  /acls/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    put:
      tags: [Routes]
      summary: Update route
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ACL' }
            example:
              id: app.example.com
              name: App
              frontend: https-443
              match: { host: app.example.com }
              mode: terminate
              backend: app
              certificate: app.example.com
              middleware:
                ip_allow: [10.0.0.0/8]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ACL' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Routes]
      summary: Delete route
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
  /backends:
    get:
      tags: [Backends]
      summary: List backends
      responses:
        '200':
          description: Pools
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Backend' }
    post:
      tags: [Backends]
      summary: Create backend
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Backend' }
            example:
              name: App
              algorithm: round_robin
              servers:
                - url: http://127.0.0.1:3000
                  role: primary
                  weight: 1
              health:
                type: http
                path: /health
                interval: 5s
                timeout: 2s
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Backend' }
        '400': { $ref: '#/components/responses/Error' }
...
  /backends/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    put:
      tags: [Backends]
      summary: Update backend
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Backend' }
            example:
              id: app
              name: App
              algorithm: round_robin
              servers:
                - url: http://127.0.0.1:3000
                  role: primary
                  weight: 1
              health:
                type: http
                path: /health
                interval: 5s
                timeout: 2s
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Backend' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Backends]
      summary: Delete backend
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
...
  /certificates:
    get:
      tags: [Certificates]
      summary: List certificates
      responses:
        '200':
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CertificateView' }
    post:
      tags: [Certificates]
      summary: Create certificate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Certificate' }
            examples:
              http01:
                summary: Let's Encrypt HTTP-01
                value:
                  name: App
                  domains: [app.example.com]
                  challenge: http-01
              dns01:
                summary: Let's Encrypt DNS-01
                value:
                  name: Wildcard
                  domains: ["*.example.com", example.com]
                  challenge: dns-01
                  dns_provider: cloudflare
                  dns_credentials:
                    CF_DNS_API_TOKEN: your-token
              custom:
                summary: Upload PEM
                value:
                  name: Custom
                  domains: [app.example.com]
                  challenge: custom
                  cert_pem: "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n"
                  key_pem: "-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CertificateView' }
        '400': { $ref: '#/components/responses/Error' }
  /certificates/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    put:
      tags: [Certificates]
      summary: Update certificate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Certificate' }
            example:
              id: app.example.com
              name: App
              domains: [app.example.com]
              challenge: http-01
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CertificateView' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Certificates]
      summary: Delete certificate
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
  /certificates/{id}/upload:
    post:
      tags: [Certificates]
      summary: Upload custom PEM
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CertUpload' }
            example:
              cert_pem: "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n"
              key_pem: "-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n"
      responses:
        '200':
          description: Uploaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOK' }
              example: { status: uploaded }
        '400': { $ref: '#/components/responses/Error' }
  /certificates/{id}/renew:
    post:
      tags: [Certificates]
      summary: Renew certificate
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Renewed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOK' }
              example: { status: renewed }
        '400': { $ref: '#/components/responses/Error' }
  /certificates/{id}/job:
    get:
      tags: [Certificates]
      summary: Certificate job log
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Issue or renew progress
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CertJob' }
  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
            example: { username: operator, password: change-me, role: operator }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Error' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/NumericID'
    put:
      tags: [Users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
            example: { role: viewer, password: "" }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Users]
      summary: Delete user
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
  /apikeys:
    get:
      tags: [API keys]
      summary: List API keys
      responses:
        '200':
          description: Keys (token is never listed)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/APIKey' }
    post:
      tags: [API keys]
      summary: Create API key
      description: The raw token is returned once in `token`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/APIKeyCreate' }
            example:
              name: ci
              scopes: [frontends:read, backends:read, acls:read, stats:read]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIKeyCreated' }
        '400': { $ref: '#/components/responses/Error' }
  /apikeys/{id}:
    delete:
      tags: [API keys]
      summary: Delete API key
      parameters:
        - $ref: '#/components/parameters/NumericID'
      responses:
        '204': { $ref: '#/components/responses/Deleted' }
        '400': { $ref: '#/components/responses/Error' }
  /audit:
    get:
      tags: [Audit]
      summary: Audit log
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
          example: 50
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
          example: 0
      responses:
        '200':
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }
...
  /config:
    get:
      tags: [Config]
      summary: Full proxy config
      responses:
        '200':
          description: Current proxy.yaml as JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyConfig' }
    put:
      tags: [Config]
      summary: Replace proxy config
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProxyConfig' }
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyConfig' }
        '400': { $ref: '#/components/responses/Error' }
  /config/validate:
    post:
      tags: [Config]
      summary: Dry-run validate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProxyConfig' }
      responses:
        '200':
          description: Valid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOK' }
        '400': { $ref: '#/components/responses/Error' }
  /config/sync:
    post:
      tags: [Config]
      summary: Pull Git and apply
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GitSync' }
        '400': { $ref: '#/components/responses/Error' }
...
  /settings:
    get:
      tags: [Config]
      summary: App settings
      description: Sanitized process settings from config.yaml. Secrets are omitted.
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
    put:
      tags: [Config]
      summary: Update app settings
      description: Updates acme_email and Git sync settings in config.yaml. Other fields are read-only here.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettingsUpdate' }
            example:
              acme_email: admin@example.com
              git:
                enabled: true
                url: git@github.com:org/goproxy-config.git
                branch: main
                auth: ssh
                key_path: /etc/goproxy/deploy_key
      responses:
        '200':
          description: Saved settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
        '400': { $ref: '#/components/responses/Error' }
  /notifications:
    get:
      tags: [Notifications]
      summary: List webhooks
      responses:
        '200':
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Webhook' }
    put:
      tags: [Notifications]
      summary: Replace webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/Webhook' }
            example:
              - url: https://hooks.example.com/goproxy
                format: generic
                triggers: [backend.down, cert.renew.failure]
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Webhook' }
        '400': { $ref: '#/components/responses/Error' }
  /notifications/test:
    post:
      tags: [Notifications]
      summary: Send a test webhook
      description: Does not persist the webhook. Works with an unsaved URL.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Webhook' }
            example:
              url: https://hooks.example.com/goproxy
              format: generic
      responses:
        '200':
          description: Delivered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOK' }
        '502': { $ref: '#/components/responses/Error' }